Implement Geocoder client
