+++ package.json
@@ -33,6 +33,7 @@
         "@turf/bbox": "^6.3.0",
         "@turf/destination": "^6.3.0",
         "@turf/helpers": "^6.3.0",
+        "@types/geojson": "^7946.0.7",
         "@types/node-fetch": "^2.5.10",
         "clean-deep": "^3.4.0",
         "node-fetch": "^2.6.1",
