+++ src/index.ts
@@ -9,6 +9,8 @@ export {
     throttler,
 } from './utils'
 
+export * as GeocoderTypes from './geocoder/types'
+
 export * as MobilityTypes from './mobility/types'
 export type { MobilityClient } from './mobility'
 
@@ -41,8 +43,11 @@ export * from './constants/featureCategory'
 export { journeyPlannerQuery, nsrQuery } from './api'
 
 export type { DeparturesById } from './departure'
-export { County } from './geocoder/countyIds'
-export type { GetFeaturesParams, GetFeaturesReverseParam } from './geocoder'
+export { County } from './geocoderLegacy/countyIds'
+export type {
+    GetFeaturesParams,
+    GetFeaturesReverseParam,
+} from './geocoderLegacy'
 
 export default createEnturService
